vector_store/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
faiss-cpu==1.7.4
numpy==1.25.0
pydantic==1.10.9
aiohttp==3.8.4